More robust installation script
(should be improved - it's a bit complex now)

+++ install.sh
@@ -6,12 +6,19 @@ DIR=$(pwd)
 cd $DIR/dotfiles/
 for i in *
 do
-    ln -s $DIR/dotfiles/$i ~/.$i
+    if [ -e ~/.$i ]; then
+      echo Will not link ~/.$i since it already exists.
+    else
+      echo "linking $i -> ~/.$i"
+      ln -s $DIR/dotfiles/$i ~/.$i
+    fi
 done
 
 # Add link to ~/bin/
-ln -s $DIR/bin/ ~/bin
+if [ ! -e ~/bin ]; then
+    ln -s $DIR/bin/ ~/bin
+fi
 
-echo "git push" > $DIR/.git/hooks/post-commit
-chmod 755 $DIR/.git/hooks/post-commit
+#echo "git push" > $DIR/.git/hooks/post-commit
+#chmod 755 $DIR/.git/hooks/post-commit
